Report SCM Polling errors as "BUILD_NOW"
The engine catches throwed exceptions and take them as "No Changes", so it's better to answer the polling request as "BUILD_NOW" (ie: "could not  check so you better try to update and build, just in case").

In this way, if the error was caused by bad credentials or any communication error, the update will probably also have the same error, and the build will be properly marked as failure, allowing the users to know about it.