Fixed ToolHelper for default installation cases
When no tool installation is specified, the received tool path is empty or null. In that case, it now defaults to the execName (which will rely on the system finding it in the workspace directory or the environment path). Before this fix, it was wrongly constructing the path as "\{execName}"